% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poc_get_api.R
\name{collect_trams_from_api}
\alias{collect_trams_from_api}
\title{Make single api call for trams geolocalisation}
\usage{
collect_trams_from_api(
  api_key = NULL,
  line_number = NULL,
  time_interval = 60L,
  max_calls = NULL,
  output = "./data-raw/api_tram_out.tsv"
)
}
\arguments{
\item{api_key}{personal key to API}

\item{tram_number}{number of a tram to track; default NULL call for all trams}
}
\value{
single call transforms JSON to tibble and saves to .tsv file
}
\description{
Make single api call for trams geolocalisation
}
