% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_functions.R
\name{get_all_lines_at_stop}
\alias{get_all_lines_at_stop}
\title{Get all lines available for particular stop}
\usage{
get_all_lines_at_stop(
  unique_stop_instances,
  max = NULL,
  api_key = Sys.getenv("UM_WAW_PERSONAL_API_KEY")
)
}
\arguments{
\item{unique_stop_instances}{dataframe with unique set of team and pole
combination}

\item{max}{max api calls, NULL makes as much calls as combinations in
unique_stop_instances}

\item{api_key}{personal api key}
}
\value{

}
\description{
Get all lines available for particular stop
}
