% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_functions.R
\name{get_all_stops_gps}
\alias{get_all_stops_gps}
\title{Get GPS coordinates for all stops in WWA}
\usage{
get_all_stops_gps(api_key = Sys.getenv("UM_WAW_PERSONAL_API_KEY"))
}
\arguments{
\item{api_key}{personal api key}
}
\description{
Get GPS coordinates for all stops in WWA
}
