% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_functions.R
\name{get_single_timetable}
\alias{get_single_timetable}
\title{Get timetable for single stop and line}
\usage{
get_single_timetable(url)
}
\arguments{
\item{url}{single api request}
}
\value{
single dataframe with timetable
}
\description{
Get timetable for single stop and line
}
