% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_functions.R
\name{get_timetables}
\alias{get_timetables}
\title{Get timetables for many stops and lines}
\usage{
get_timetables(
  unique_stop_instances,
  lines,
  api_key = Sys.getenv("UM_WAW_PERSONAL_API_KEY")
)
}
\arguments{
\item{unique_stop_instances}{unique set of team and pole}

\item{lines}{lines}

\item{api_key}{personal api key}
}
\value{
nested dataframe with timetables
}
\description{
Get timetables for many stops and lines
}
