% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_functions.R
\name{collect_trams_from_api}
\alias{collect_trams_from_api}
\title{Make single api call for ?trams geolocalisation}
\usage{
collect_trams_from_api(
  line_number = NULL,
  time_interval = 60L,
  max_calls = NULL,
  output = "./data-raw/api_tram_out.tsv",
  api_key = Sys.getenv("UM_WAW_PERSONAL_API_KEY")
)
}
\arguments{
\item{line_number}{tram line number to call, default NULL will call all trams}

\item{time_interval}{time between API calls in seconds}

\item{max_calls}{maximum number of calls to make}

\item{output}{output file path and name}

\item{api_key}{personal key to API}
}
\value{
single call transforms JSON to tibble and saves to .tsv file
}
\description{
Make single api call for ?trams geolocalisation
}
